Drop ignored properties from task definition
* `compatibilities`
* `requiresAtttributes`
* `revision`
* `status`

Reported as warnings in Actions annotations:

Ignoring property XXX in the task definition file. This
property is returned by the Amazon ECS DescribeTaskDefinition API and
may be shown in the ECS console, but it is not a valid field when
registering a new task definition. This field can be safely removed from
your task definition file.