PO Line # from GET response, not internal map #32
The follow up GET to `/orders/composite-orders/{id}` includes both the
`poLineNumber` and `instanceId` per PO Line in the response [1].

This ensures we report back the correct instance associated with each PO
Line in the results block rendered to the user. All that remains is to
now correctly report the instance Title as well.

[1]: https://s3.amazonaws.com/foliodocs/api/mod-orders/p/order.html#orders_composite_orders__id__get