Use instance title from composite order GET #32
Ensures POL and Instance reported in feedback to user are completely in
sync. Hat tip to @fereira for pointing out the `titleOrPackage` key in
the response.

He plans on refactoring this in a followup PR, creating an interator for
the `updatedPurchaseOrderJson.compositePoLines` JSONArray in lieu of
iterating through the records in the submitted MARC file a second time.