Urlencode cql query for organization lookup
Forward slash is a special character in CQL [1], so urlencode to avoid
`org.folio.cql2pgjson.exception.QueryValidationException` and support
Cornell LTS' heavy use of organization (vendor) codes with forward
slashes.

We do plan to consolidate the organizations and eliminate codes with
forward slashes but this will not be ready for go-live.

Also note that the FOLIO Organizations app (ui-organizations) does not
urlencode the organization code in the edit form, and this does nothing
to address that bug. In oother words, we still will be unable to edit
organizations via the UI which have codes that contain forward slashes.

One final note. The validateRequiredValues() is ripe for refactoring so
we can avoid redundant HTTP requests. See TODO.

[1]: https://www.loc.gov/standards/sru/cql/spec.html