Set quantity (980$g) for cost & locations
The composite order endpoint tracks quantity as `quantityPhysical` under
cost and locations objects. Previously we were hardcoding the value
under cost to `1`, while dynamically setting the value under locations
based on the `980$g` in the MARC record.

Now, both are set dynamically via the `980$g` and remain in sync, thus
avoiding the `physicalLocCostQtyMismatch` error when POSTing the
composite order.

> "PO Line physical quantity and Locations physical quantity do not
> match"